Move workOnTypes to its own module
This removes a dependency cycle between Reduce and workOnTypes, meaning
we no longer need to SOURCE import. This in turn allows workOnTypes to
be inlined/specialised, avoiding a common cause of allocations.

This provides a ~2% speedup in typechecking.
