+++ src/full/Agda/TypeChecking/Coverage.hs
@@ -63,6 +63,7 @@ import Agda.TypeChecking.Telescope
 import Agda.TypeChecking.Telescope.Path
 import Agda.TypeChecking.MetaVars
 import Agda.TypeChecking.Warnings
+import Agda.TypeChecking.WorkOnTypes
 
 import Agda.Interaction.Options
