+++ src/full/Agda/TypeChecking/Rules/Record.hs
@@ -28,7 +28,7 @@ import Agda.TypeChecking.Positivity.Occurrence
 import Agda.TypeChecking.Pretty
 import Agda.TypeChecking.Polarity
 import Agda.TypeChecking.Warnings
-import Agda.TypeChecking.Irrelevance
+import Agda.TypeChecking.WorkOnTypes
 import Agda.TypeChecking.CompiledClause (hasProjectionPatterns)
 import Agda.TypeChecking.CompiledClause.Compile
